# backend/src/main/java/com/leonardo/hug/interaction/InteractionRequest.java
package com.leonardo.hug.interaction;

import jakarta.persistence.EnumType;
import jakarta.persistence.Enumerated;

public record InteractionRequest(
    long productId,
    @Enumerated(EnumType.STRING)
    InteractionType interactionType,
    double duration
) {
    
}